创建客户的同时添加联系人

+++ app/entitys/crm_linkman.go
@@ -4,7 +4,7 @@ import "time"
 
 type CrmLinkman struct {
 	Id         int       `orm:"column(id);auto"`
-	CustId     int       `orm:"column(cust_id);null" description:"客户id"`
+	CustId     int64       `orm:"column(cust_id);null" description:"客户id"`
 	Name       string    `orm:"column(name);size(255);null" description:"联系人姓名"`
 	Job        string    `orm:"column(job);size(255);null" description:"职业"`
 	Call       string    `orm:"column(call);size(255);null" description:"固定电话"`
@@ -13,7 +13,7 @@ type CrmLinkman struct {
 	Wechat     string    `orm:"column(wechat);size(256);null" description:"微信"`
 	Email      string    `orm:"column(email);size(255);null" description:"邮箱"`
 	Main       string    `orm:"column(main);null"`
-	Sex        int8      `orm:"column(sex);null"`
+	Sex        int      `orm:"column(sex);null"`
 	Address    string    `orm:"column(address);size(255);null"`
 	Intro      string    `orm:"column(intro);size(255);null"`
 	CreateTime time.Time `orm:"column(create_time);type(timestamp);auto_now_add" description:"创建时间"`
